Enable native-bootstrap to continue build after recovered from fault.
The previous script creates new directory for the chunk it is going
to build without checking whether the directory exists or not.
It will fail back if the directory it attempted to creat exists.

So, you always need to make sure no .inst directory exists and let
the native-bootstrap script build from the beginning. It's time consuming
if any error happens, and you need to continue build after recoverred.

This patch improves this, and you can run the native bootstrap script direct
after resumed.

Change-Id: I91ae213ecc8c98808efdfd969624291e70f7e0fe