rawdisk.write: Add a function to find device names for partitions
This function finds the name assigned by the system to real block
devices with partitions on them, which is not always done in the same
way. This function finds the name given to a numbered partition.

E.g. Return '/dev/sda1' when asked for the first partition on '/dev/sda'
     Return '/dev/mmcblk1p1' for the first partition on '/dev/mmcblk1'